Extract logging tests for the cause handling into a separate file
