+++ tests/handling/test_cause_handling.py
@@ -1,28 +1,11 @@
 import asyncio
 import logging
 
-import pytest
-
 import kopf
-from kopf.reactor.causation import ALL_CAUSES, CREATE, UPDATE, DELETE, NEW, GONE, FREE, NOOP
+from kopf.reactor.causation import CREATE, UPDATE, DELETE, NEW, GONE, FREE, NOOP
 from kopf.reactor.handling import custom_object_handler
 from kopf.structs.finalizers import FINALIZER
 from kopf.structs.lastseen import LAST_SEEN_ANNOTATION
-
-
-@pytest.mark.parametrize('cause_type', ALL_CAUSES)
-async def test_all_logs_are_prefixed(registry, resource, caplog, cause_type, cause_mock):
-    caplog.set_level(logging.DEBUG)
-    cause_mock.event = cause_type
-
-    await custom_object_handler(
-        lifecycle=kopf.lifecycles.all_at_once,
-        registry=registry,
-        resource=resource,
-        event={'type': 'irrelevant', 'object': cause_mock.body},
-        freeze=asyncio.Event(),
-    )
-    assert all(message.startswith('[ns1/name1] ') for message in caplog.messages)
 
 
 async def test_new(registry, handlers, resource, cause_mock,
